+++ Manager/Resources/Changelog.md
@@ -3,6 +3,7 @@ Modnix Changelog
 # Version 3 Beta 2, in development
 Roadmap: https://github.com/Sheep-y/Modnix/issues/24
 
+* Fix: Create app config on setup to allow "remote" dll load.
 * Fix: Try to make sandbox assembly path local.
 * Log: OS type and version are now logged.